update OTA mDNS and add soft reset config wifi

+++ Timmer.h
@@ -9,7 +9,7 @@ class Timmer
 public:
     int relayID = -1;   
     int relayConditionNumber;  // Điều kiện thứ i của relay.listCondigion
-    int timmerCycle;    // chu trình vòng lặp 
+    long timmerCycle;    // chu trình vòng lặp 
     long timmerStart = 0;;  // Thời gian bắt đầu
     long timmerEnd = 0;;    // Thời gian kết thúc
     int timmerInfluence = RELAY_OFF;   // Tác động lên relay
